==> 3a-qc-methylation.Rmd <==
# QC analysis of methylation data

```{r setup, include=FALSE}
# source("functions/knit_wiki.r")
# knit_wiki("3a-qc-methylation.Rmd", output.filename = "Methylation data QC analysis")

report.title <- "qc-methylation"
fig.dir <- file.path("figures", report.title)
fig.prefix <- "meth-"

dir.create(fig.dir, showWarnings = FALSE)

opts_knit$set(upload.fun = function(file, 
 key = "fbd62838790b8f6192669e30c3fb5346") imgur_upload(file, key))

opts_chunk$set(fig.path = file.path(fig.dir, fig.prefix),
  cache.path = paste(".cache-", report.title, "/", sep = ""),
  echo=FALSE, results='hide',
  cache=TRUE, autodep=TRUE, fig.width=8.5, fig.height=5, dpi=150)
```

```{r load-libraries, message=FALSE, cache=FALSE}
library(affy)
library(ggplot2)
library(plyr)
library(ggaffy)

source("functions/print_xtable.r")
source("functions/save-figures.r")
```

```{r graphing-tweaks}
# Shrink text so labels are legible along x-axis
axis.text.x <- theme_text(size = 5, colour = "grey50", angle = 90, hjust = 1)

wide.opts <- opts(axis.text.x = axis.text.x, legend.position = "top",
  legend.margin = grid::unit(0, "lines"))
```


```{r load-data, cache=FALSE}
load("data/eset-meth.rda")
```

## Sample filters

### Count NAs per sample
The following histogram depicts the number of NA values per sample. While most samples contain on average about ~16 missing values, there is a subset of samples with more than 50 missing values. 

```{r na-count-histogram, message=FALSE, fig.width=5, fig.height=3}
eset.meth$na.count <- apply(exprs(eset.meth), 2, function(x) sum(is.na(x)))

qplot(eset.meth$na.count) + 
  ylab("Sample count") + xlab("NA methylation probe count")
```

```{r na-count-individuals}
ggplot(pData(eset.meth)) + 
  aes(x = individual, y = na.count, fill = batch) +
  geom_bar() + facet_wrap(~tissue, ncol = 1) +
  ylab("NA methylation probe count") +
  scale_fill_brewer(palette = "Set1") +
  opts(legend.position = "top", axis.text.x = axis.text.x)
```

### Count NAs per probe
A few red probes stand out as having a high number of missing values.

```{r na-count-probes, fig.width=5, fig.height=5}
fData(eset.meth)$na.count <- apply(exprs(eset.meth), 1, 
  function(x) sum(is.na(x)))

qplot(data = fData(eset.meth), x =  na.count, geom = "density",
      facets = color.channel ~ ., fill = I("black")) + geom_rug()
```

```{r na-count-probes-table, results='asis'}
print_xtable(arrange(fData(eset.meth), 
  desc(na.count))[1:10, c("id", "symbol", "na.count")])
```


## Distributions of unnormalized data

### Quantiles plot
```{r quantiles}
ggaffy_quantiles(eset.meth,
  x = "individual", probs = c(0.01, 0.1, 0.25),
  fill = "batch", shape = "gender", facet.row = "tissue") +
  wide.opts
```

### Boxplot
```{r boxplot}
ggaffy_box(eset.meth, x = "individual", fill = "batch", facet.row = "tissue") +
  wide.opts
```

### Relative expression boxplot
```{r boxplot-rle}
ggaffy_box(eset.meth, x = "individual", method = "rle",
  fill = "batch", facet.row = "tissue") + wide.opts
```

### Histogram
```{r histogram, fig.width=5, fig.height=8.5}
ggaffy_hist(eset.meth, facet.row = "tissue", color = "batch", adjust = 0.5) +
  wide.opts
```

## Sample clustering
Hierarchical clustering of all samples using euclidean distance produces results very similar to what Gibbs published in [figure 1a](http://www.plosgenetics.org/article/info:doi/10.1371/journal.pgen.1000952?imageURI=info:doi/10.1371/journal.pgen.1000952.g001) of their paper.

```{r sample-cluster-heatmap, fig.height=4, fig.width=4.5}
ggaffy_clustmat(eset.meth, color = "tissue", raster = TRUE, 
  dist.method = "euclidean") + 
opts(axis.text.x = theme_blank(), axis.text.y = theme_blank(),
  axis.ticks = theme_blank())
```

## Sample comparisons with median reference
Inspect the relationship between each sample and a pseudo-reference sample, generated from the gene-wise median across all samples within a given tissue. The following table lists all samples with squared median deviations that are at least 2 SDs above the tissue-specific mean. 

```{r median-ref-comparisons, results='asis'}
tissues <- levels(eset.meth$tissue)

# Consider only autosomal probes
eset.auto <- eset.meth[fData(eset.meth)$chr %in% 1:22, ]

# Reference comparison results
ref.comps <- list()
for(t in tissues) {
  eset.t <- eset.auto[, eset.auto$tissue == t]
  exp.mat <- exprs(eset.t)
  med.ref <- rowMedians(exp.mat)
  
  # Correlations with median
  cors.sub <- apply(exp.mat, 2, cor, y = med.ref, use = "pairwise.complete")
  
  # Sum of deviations from the median
  devs.sub <- apply(exp.mat, 2, function(x) sum((med.ref - x)^2, na.rm = TRUE))
  
  out <- data.frame(tissue = t, sample = sampleNames(eset.t),
    individual = eset.t$individual, cor = cors.sub, dev = devs.sub)
  ref.comps[[t]] <- out
}

ref.comps <- data.frame(do.call("rbind", ref.comps), row.names = NULL)

# Calculate outlier threshold
thresh <- mean(ref.comps$dev) + (sd(ref.comps$dev) * 2)

# Table of outlier samples
ref.comps$outlier <- ifelse(ref.comps$dev > thresh, TRUE, FALSE)
ref.comps <- arrange(ref.comps, desc(dev))
print_xtable(subset(ref.comps, outlier, select = -outlier))

# Export comparisons
write.csv(ref.comps, paste(file.path(fig.dir, fig.prefix), 
  "median-ref-comparisons.csv", sep = ""), row.names = FALSE)
```

### Outlier threshold
Below is a summary of all sample deviations from the median reference, the red line indicates the outlier threshold. 
```{r median-ref-deviations, fig.width=3, fig.height=5}
qplot(data = ref.comps, x = factor(1), y = dev, geom = "boxplot") +
  geom_hline(yintercept = thresh, color = "red") +
  ylab("Sum of squared median deviations") +
  opts(axis.text.x = theme_blank(), axis.title.x = theme_blank())
```


### Cerebellum median reference comparison
```{r median-ref-comparisons-crblm, fig.height = 8.5}
eset.t <- eset.auto[, eset.auto$tissue == "CRBLM"]
sampleNames(eset.t) <- eset.t$individual
ggaffy_scatgrid(eset.t)
```

### Pons median reference comparison
```{r median-ref-comparisons-pons, fig.height = 8.5}
eset.t <- eset.auto[, eset.auto$tissue == "PONS"]
sampleNames(eset.t) <- eset.t$individual
ggaffy_scatgrid(eset.t)
```

### Frontal cortex median reference comparison
```{r median-ref-comparisons-fctx, fig.height = 8.5}
eset.t <- eset.auto[, eset.auto$tissue == "FCTX"]
sampleNames(eset.t) <- eset.t$individual
ggaffy_scatgrid(eset.t)
```

### Temporal cortex median reference comparison
```{r median-ref-comparisons-tctx, fig.height = 8.5}
eset.t <- eset.auto[, eset.auto$tissue == "TCTX"]
sampleNames(eset.t) <- eset.t$individual
ggaffy_scatgrid(eset.t)
```
